backend/requirements.txt
# Flask & Extensions
Flask>=3.0.0,<4.0.0
Flask-CORS>=4.0.0,<6.0.0
Flask-JWT-Extended>=4.6.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0

# Production Server
gunicorn>=21.0.0,<23.0.0

# Database
pymongo>=4.6.0,<5.0.0
certifi>=2023.0.0
dnspython>=2.4.0,<3.0.0

# Authentication
bcrypt>=4.1.0,<5.0.0

# AI & Processing (lightweight - no EasyOCR/Tesseract for Render free tier)
# Using PyMuPDF text extraction instead of heavy OCR
nltk>=3.8.0,<4.0.0
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0
rapidfuzz>=3.5.0,<4.0.0

# PDF Processing
PyMuPDF>=1.23.0,<2.0.0

# Utilities
Werkzeug>=3.0.0,<4.0.0
Pillow>=10.0.0,<12.0.0

# Optional: Memory monitoring (for health checks)
psutil>=5.9.0,<6.0.0

# Document Support
python-docx>=1.1.0,<2.0.0

# Tunneling (for public access from Colab / local dev)
pyngrok>=7.0.0,<8.0.0
